+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix Maven installation script
 - Fix SIP integration test
 - Fixed v1 meeting bug related to bootstrap row class
+- Fix meeting readiness checker integration test
 
 ## [1.18.0] - 2020-09-22
 ### Added